[Parser] Diagnostics now work in the parser
