+++ src/ir_parser.c
@@ -8,11 +8,7 @@
 #include <stdarg.h>
 #include <string.h>
 
-/// For isatty().
 #ifdef _WIN32
-#    include <io.h>
-#    define isatty _isatty
-
 /// Basically strndup(), but for Windows.
 char *strndup(const char *str, size_t sz) {
   char *dup = malloc(sz + 1);
@@ -20,142 +16,7 @@ char *strndup(const char *str, size_t sz) {
   dup[sz] = 0;
   return dup;
 }
-
-#else
-#    include <unistd.h>
 #endif
-
-
-typedef uint32_t u32;
-typedef uint64_t u64;
-typedef int32_t i32;
-typedef int64_t i64;
-typedef size_t usz;
-
-/// ===========================================================================
-///  Diagnostics
-/// ===========================================================================
-enum diagnostic_level {
-  DIAG_NOTE,
-  DIAG_WARN,
-  DIAG_ERR,
-  DIAG_ICE,
-  DIAG_SORRY,
-
-  DIAG_COUNT,
-};
-
-/// Source location.
-typedef struct {
-  u32 start;
-  u32 end;
-} loc;
-
-static const char *diagnostic_level_names[DIAG_COUNT] = {
-  "Note",
-  "Warning",
-  "Error",
-  "Internal Compiler Error",
-  "Sorry, unimplemented",
-};
-
-static const char *diagnostic_level_colours[DIAG_COUNT] = {
-  "\033[1;34m",
-  "\033[1;36m",
-  "\033[1;31m",
-  "\033[1;31m",
-  "\033[1;31m",
-};
-
-/// Issue a compiler diagnostic.
-///
-/// WARNING: ALTER THIS FUNCTION AT YOUR OWN PERIL.
-///
-/// This function is the culmination of many, *many* hours of battling
-/// off-by-one errors and banging my head against the nearest wall. It
-/// was copied from a previous project because I tried reproducing it
-/// at first, but failed horribly. – Sirraide.
-FORMAT(printf, 5, 6)
-  void issue_diagnostic(
-  /// Error level and source location.
-  enum diagnostic_level level,
-  const char *filename,
-  span source,
-  loc location,
-
-  /// The actual error message.
-  const char *fmt,
-  ...) {
-  ASSERT(level >= 0 && level < DIAG_COUNT);
-
-  /// Check if stderr is a terminal.
-  bool is_terminal = isatty(fileno(stderr));
-
-  if (location.start > source.size) location.start = (u32) (source.size);
-  if (location.end > source.size) location.end = (u32) (source.size);
-
-  /// Seek to the start of the line. Keep track of the line number.
-  u32 line = 1;
-  u32 line_start = 0;
-  for (u32 i = location.start; i > 0; --i) {
-    if (source.data[i] == '\n') {
-      if (!line_start) line_start = i + 1;
-      ++line;
-    }
-  }
-
-  /// Don’t include the newline in the line.
-  if (source.data[line_start] == '\n') ++line_start;
-
-  /// Seek to the end of the line.
-  u32 line_end = location.end;
-  while (line_end < source.size && source.data[line_end] != '\n') line_end++;
-
-  /// Print the filename, line and column, severity and message.
-  if (is_terminal) fprintf(stderr, "\033[m\033[1;38m");
-  fprintf(stderr, "%s:%u:%u: ", filename, line, location.start - line_start);
-  if (is_terminal) fprintf(stderr, "%s", diagnostic_level_colours[level]);
-  fprintf(stderr, "%s: ", diagnostic_level_names[level]);
-  if (is_terminal) fprintf(stderr, "\033[m\033[1;38m");
-  va_list args;
-  va_start(args, fmt);
-  vfprintf(stderr, fmt, args);
-  va_end(args);
-  if (is_terminal) fprintf(stderr, "\033[m");
-
-  /// Print the line.
-  fprintf(stderr, "\n %d | ", line);
-  for (u32 i = line_start; i < location.start; ++i) {
-    if (source.data[i] == '\t') fprintf(stderr, "    ");
-    else fputc(source.data[i], stderr);
-  }
-  if (is_terminal) fprintf(stderr, "%s", diagnostic_level_colours[level]);
-  for (u32 i = location.start; i < location.end; ++i) {
-    if (source.data[i] == '\t') fprintf(stderr, "    ");
-    else fputc(source.data[i], stderr);
-  }
-  if (is_terminal) fprintf(stderr, "\033[m");
-  for (u32 i = location.end; i < line_end; ++i) {
-    if (source.data[i] == '\t') fprintf(stderr, "    ");
-    else fputc(source.data[i], stderr);
-  }
-  fprintf(stderr, "\n");
-
-  /// Underline the region with tildes.
-  size_t spaces = !line ? 1 : (u32) (log10(line) + 1);
-  for (size_t i = 0; i < spaces; i++)
-    fprintf(stderr, " ");
-  fprintf(stderr, "  | %s", is_terminal ? diagnostic_level_colours[level] : "");
-  for (u32 i = line_start; i < location.start; ++i) {
-    if (source.data[i] == '\t') fprintf(stderr, "    ");
-    else fputc(' ', stderr);
-  }
-  for (u32 i = location.start; i < location.end; ++i) {
-    if (source.data[i] == '\t') fprintf(stderr, "~~~~");
-    else fputc('~', stderr);
-  }
-  if (is_terminal) fprintf(stderr, "\033[m\n");
-}
 
 /// ===========================================================================
 ///  Lexer
